+++ v2/test/v2Tests/index.js
@@ -24,7 +24,6 @@ module.exports = function() {
   describe("paleogeography", require("./paleogeography"));
   describe("geologic_units/gmna", require("./geologic_units_gmna"));
   describe("geologic_units/gmus", require("./geologic_units_gmus"));
-  describe("geologic_units/intersection", require("./geologic_units_intersection"));
 
   describe("mobile/point", require("./mobile_point"));
   describe("mobile/point_details", require("./mobile_point_details"));